cloud-function/requirements.txt
google-cloud-storage==2.10.0
functions-framework==3.4.0
flask==2.3.3
